refactor: move auth api middleware to auth folder
